+++ CHANGES.md
@@ -25,7 +25,7 @@ Changes:
 Fixes:
 - `conn` and `connappct` now use ClusterSite #1733
 - fix `tenn` miss adding opinion types #1749
-
+- add ADDENDUM to OpinionType enum #1754
 
 ## Current